Make kubeadm join idempotent

+++ pkg/installer/version/kube112/05-master-join-cluster.go
@@ -17,6 +17,7 @@ func joinMasterCluster(ctx *util.Context) error {
 }
 func joinControlplaneNode(ctx *util.Context, node *config.HostConfig, conn ssh.Connection) error {
 	_, _, _, err := util.RunShellCommand(conn, ctx.Verbose, `
+if [[ -f /etc/kubernetes/kubelet.conf ]]; then exit 0; fi
 if [[ -f /etc/systemd/system/kubelet.service.d/10-kubeadm.conf.disabled ]]; then
 	sudo mv /etc/systemd/system/kubelet.service.d/10-kubeadm.conf{.disabled,}
 	sudo systemctl daemon-reload
